+++ src/main/java/me/preetham/samsaram/repository/BankRepository.java
@@ -3,5 +3,5 @@ package me.preetham.samsaram.repository;
 import me.preetham.samsaram.model.Bank;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BankRepository extends CrudRepository<Bank, Long> {
+public interface BankRepository extends CrudRepository<Bank, Integer> {
 }
